12_OpenAI_Agent_SDK/12_Sessions/pyproject.toml
[project]
name = "12-sessions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
